third_party/blink/web_tests/platform/generic/netinfo/gc-frame-listeners-expected.txt

Tests that listeners in closed frames are collected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS childFrameObserver.wasCollected is false
PASS childConnectionObserver.wasCollected is false
PASS callbackObserver.wasCollected is false
PASS childFrameObserver.wasCollected is true
PASS childConnectionObserver.wasCollected is true
PASS callbackObserver.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
